Fixed vendor folder conflicts. The root of the problem was that there were static files in the vendor folder (sweetalert2.js and .css to be specific) that were required for the web page. That's what generated in previous versions the symlink under the src folder.
This time the fix is structural. ALL static files should be under src/static , including those coming from a vendor library, which _will_ require the developer to move that minimized file or stylesheet to the static folder.

I have also added a default set of phpmd configurations, and some tweaks to the packaging of the build.xml file

Joe